+++ src/app/app.routes.ts
@@ -3,7 +3,6 @@ import { MessageService } from './features/messages/services/message.service';
 import { authGuard } from './features/auth/guards/auth.guard';
 import { unauthenticatedGuard } from './features/auth/guards/unauthenticated.guard';
 import { ConversationService } from './features/messages/services/conversation.service';
-// import { NotificationService } from './features/messages/services/notification.service';
 
 export const routes: Routes = [
   {
@@ -31,11 +30,7 @@ export const routes: Routes = [
       import('./features/messages/messages.routes').then(
         (m) => m.messagesRoutes,
       ),
-    providers: [
-      MessageService,
-      ConversationService,
-      // NotificationService
-    ],
+    providers: [MessageService, ConversationService],
     canActivate: [authGuard],
   },
   {